\name{seedbank}
\alias{seedbank}
\docType{data}
\title{
Seed abundances in soil samples
}
\description{
Number of seeds of different species germinating from 121 soil samples collected in a tropical secondary forest.
}
\usage{data(seedbank)}
\format{
  A matrix with 34 rows for species and 121 columns corresponding to different soil samples.
}

\source{
Butler & Chazdon 1998. Example data distributed with the EstimateS program (Colwell 2005).
}
\references{
Butler, B. J., & R. L. Chazdon. 1998. Species richness, spatial variation, and abundance of the soil seed bank of a secondary tropical rain forest. \emph{Biotropica} 30:214-222.

Colwell, R K; J A Coddington. 1994. Estimating terrestrial biodiversity through extrapolation. \emph{Philosophical Transactions of the Royal Society of London B} 345:101-118.

Colwell, R. K. 2005. EstimateS: Statistical estimation of species richness and shared species from samples. Version 7.5. User's Guide and application published at: http://purl.oclc.org/estimates.
}
\examples{
data(seedbank)
## 
}
\keyword{datasets}
