\name{deprecated}
\alias{plotPost}
\title{
Deprecated functions in package wiqid
}
\description{
These functions have been replaced by functions in \code{mcmcOutput}
}
\usage{
plotPost(...)
}
\arguments{
  \item{\dots}{
arguments passed to the new function.
}
}
\details{
plotPost has been replaced by postPlot.
}
\value{
See help for the new function.
}
\author{
Mike Meredith
}
