\name{Renamed}
\alias{renamed}
\alias{occSScov}
\alias{occMSseason}
\alias{occMScov}

\title{
Functions which have been renamed
}
\description{
\code{occSScov} is replaced by \code{occSS}

\code{occMScov} is replaced by \code{occMS}

\code{occMSseason} is replaced by \code{occMStime}
}
\usage{
occSScov(DH=NULL, model=NULL, data = NULL, ci=0.95)

occMSseason(DH=NULL, occsPerSeason=NULL, model=NULL,
             data=NULL, ci=0.95)

occMScov(DH=NULL, occsPerSeason=NULL, model=NULL,
             data=NULL, ci=0.95)
             
}
\arguments{
 \item{DH}{
}
  \item{occsPerSeason}{
}
  \item{model}{
}
  \item{data}{
}
  \item{ci}{
}
}
\details{
All these functions display text with the new function name.
}
\value{
None.
}

