\name{View}
\alias{View}
\alias{View.Bwiqid}
\title{
View an object in a spreadsheet-style window
}
\description{
Invoke a spreadsheet-style data viewer on an R object. This is an S3 generic version of \code{\link[utils]{View}}.
}
\usage{
View(x, title, ...)

\method{View}{Bwiqid}(x, title, digits=4, ...)  

}
\arguments{
  \item{x}{
the object to be displayed
}
  \item{title}{
title for the viewer window.
}
  \item{\dots}{
further arguments passed to or from other methods.
}
  \item{digits}{
the number of digits to display after the decimal.
}
}


