\name{Renamed}
\alias{renamed}
\alias{occSScov}
\alias{occMSseason}
\alias{occMScov}

\title{
Functions which have been renamed
}
\description{
\code{occSScov} is replaced by \code{occSS}

\code{occMScov} is replaced by \code{occMS}

\code{occMSseason} is replaced by \code{occMStime}
}
\usage{
occSScov(...)

occMSseason(...)

occMScov(...)
             
}
\arguments{
 \item{\dots}{Arguments passed to the new functions.
}
 }
\details{
All these functions display text with the new function name.
}
\value{
None.
}

